Fix to use standard SQL type.

+++ net.solarnetwork.node.hw.loxone/src/net/solarnetwork/node/loxone/dao/jdbc/vevent-init.sql
@@ -3,7 +3,7 @@ CREATE TABLE solarnode.loxone_vevent (
 	uuid_lo		BIGINT NOT NULL,
 	config_id	BIGINT NOT NULL,
 	created		TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	fvalue      DOUBLE NOT NULL,
+	fvalue      DOUBLE PRECISION NOT NULL,
 	CONSTRAINT loxone_vevent_pk PRIMARY KEY (config_id, uuid_hi, uuid_lo)
 );
